# Use tmpfs for running benchmarks