--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -724,6 +724,7 @@ jobs:
           cd ..
           make -C test/data 
       - run: du -sh base/build pr/build
+      - run: mkdir -p /opt/benchmarks
       - name: Run PR Benchmarks 
         run: |
           sudo mount -t tmpfs -o size=4g none /opt/benchmarks