--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -714,15 +714,15 @@ jobs:
           make -j$(nproc) benchmarks
           cd ..
           make -C test/data 
-      - name: Build PR Branch
-        run: |
-          mkdir -p pr/build
-          cd pr/build
-          cmake -DENABLE_CONAN=ON -DCMAKE_BUILD_TYPE=Release ..
-          make -j$(nproc) 
-          make -j$(nproc) benchmarks
-          cd ..
-          make -C test/data 
+      # - name: Build PR Branch
+      #   run: |
+      #     mkdir -p pr/build
+      #     cd pr/build
+      #     cmake -DENABLE_CONAN=ON -DCMAKE_BUILD_TYPE=Release ..
+      #     make -j$(nproc) 
+      #     make -j$(nproc) benchmarks
+      #     cd ..
+      #     make -C test/data 
       - run: du -sh base/build pr/build
       - run: mkdir -p /opt/benchmarks
       - name: Run PR Benchmarks 
@@ -731,23 +731,24 @@ jobs:
           cp -rf pr/build /opt/benchmarks/build
           mkdir -p /opt/benchmarks/test
           cp -rf pr/test/data /opt/benchmarks/test/data
+          cp -rf pr/profiles /opt/benchmarks/profiles
           ls -R /opt/benchmarks
           ./pr/scripts/ci/run_benchmarks.sh -f /opt/benchmarks -r $(pwd)/pr_results -s $(pwd)/pr -b /opt/benchmarks/build -o ~/data.osm.pbf -g ~/gps_traces.csv
           sudo umount /opt/benchmarks
       - run: ls -R /opt/benchmarks
-      - name: Run Base Benchmarks
-        run: |
-          sudo mount -t tmpfs -o size=4g none /opt/benchmarks
-          cp -rf base/build /opt/benchmarks/build
-          mkdir -p /opt/benchmarks/test
-          cp -rf base/test/data /opt/benchmarks/test/data
-          ls -R /opt/benchmarks
-          # we intentionally use scripts from PR branch to be able to update them and see results in the same PR
-          ./pr/scripts/ci/run_benchmarks.sh -f /opt/benchmarks -r $(pwd)/base_results -s $(pwd)/pr -b /opt/benchmarks/build -o ~/data.osm.pbf -g ~/gps_traces.csv
-          sudo umount /opt/benchmarks
-      - name: Post Benchmark Results
-        run: |
-          python3 pr/scripts/ci/post_benchmark_results.py base_results pr_results
+      # - name: Run Base Benchmarks
+      #   run: |
+      #     sudo mount -t tmpfs -o size=4g none /opt/benchmarks
+      #     cp -rf base/build /opt/benchmarks/build
+      #     mkdir -p /opt/benchmarks/test
+      #     cp -rf base/test/data /opt/benchmarks/test/data
+      #     ls -R /opt/benchmarks
+      #     # we intentionally use scripts from PR branch to be able to update them and see results in the same PR
+      #     ./pr/scripts/ci/run_benchmarks.sh -f /opt/benchmarks -r $(pwd)/base_results -s $(pwd)/pr -b /opt/benchmarks/build -o ~/data.osm.pbf -g ~/gps_traces.csv
+      #     sudo umount /opt/benchmarks
+      # - name: Post Benchmark Results
+      #   run: |
+      #     python3 pr/scripts/ci/post_benchmark_results.py base_results pr_results
       - name: Show CCache statistics
         run: |
           ccache -p