--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -742,6 +742,7 @@ jobs:
       #     cp -rf base/build /opt/benchmarks/build
       #     mkdir -p /opt/benchmarks/test
       #     cp -rf base/test/data /opt/benchmarks/test/data
+      #     cp -rf base/profiles /opt/benchmarks/profiles
       #     ls -R /opt/benchmarks
       #     # we intentionally use scripts from PR branch to be able to update them and see results in the same PR
       #     ./pr/scripts/ci/run_benchmarks.sh -f /opt/benchmarks -r $(pwd)/base_results -s $(pwd)/pr -b /opt/benchmarks/build -o ~/data.osm.pbf -g ~/gps_traces.csv