--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -705,25 +705,25 @@ jobs:
         with:
           ref: ${{ github.event.pull_request.base.ref }}
           path: base
-      - name: Build Base Branch
-        run: |
-          mkdir base/build
-          cd base/build
-          cmake -DENABLE_CONAN=ON -DCMAKE_BUILD_TYPE=Release ..
-          make -j$(nproc)
-          make -j$(nproc) benchmarks
-          cd ..
-          make -C test/data 
-      # - name: Build PR Branch
+      # - name: Build Base Branch
       #   run: |
-      #     mkdir -p pr/build
-      #     cd pr/build
+      #     mkdir base/build
+      #     cd base/build
       #     cmake -DENABLE_CONAN=ON -DCMAKE_BUILD_TYPE=Release ..
-      #     make -j$(nproc) 
+      #     make -j$(nproc)
       #     make -j$(nproc) benchmarks
       #     cd ..
       #     make -C test/data 
-      # - run: du -sh base/build pr/build
+      - name: Build PR Branch
+        run: |
+          mkdir -p pr/build
+          cd pr/build
+          cmake -DENABLE_CONAN=ON -DCMAKE_BUILD_TYPE=Release ..
+          make -j$(nproc) 
+          make -j$(nproc) benchmarks
+          cd ..
+          make -C test/data 
+     # - run: du -sh base/build pr/build
       - run: mkdir -p /opt/benchmarks
       - name: Run PR Benchmarks 
         run: |