--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -705,15 +705,15 @@ jobs:
         with:
           ref: ${{ github.event.pull_request.base.ref }}
           path: base
-      # - name: Build Base Branch
-      #   run: |
-      #     mkdir base/build
-      #     cd base/build
-      #     cmake -DENABLE_CONAN=ON -DCMAKE_BUILD_TYPE=Release ..
-      #     make -j$(nproc)
-      #     make -j$(nproc) benchmarks
-      #     cd ..
-      #     make -C test/data 
+      - name: Build Base Branch
+        run: |
+          mkdir base/build
+          cd base/build
+          cmake -DENABLE_CONAN=ON -DCMAKE_BUILD_TYPE=Release ..
+          make -j$(nproc)
+          make -j$(nproc) benchmarks
+          cd ..
+          make -C test/data 
       - name: Build PR Branch
         run: |
           mkdir -p pr/build
@@ -723,7 +723,7 @@ jobs:
           make -j$(nproc) benchmarks
           cd ..
           make -C test/data 
-     # - run: du -sh base/build pr/build
+      - run: du -sh base/build pr/build
       - run: mkdir -p /opt/benchmarks
       - name: Run PR Benchmarks 
         run: |
@@ -736,20 +736,20 @@ jobs:
           ./pr/scripts/ci/run_benchmarks.sh -f /opt/benchmarks -r $(pwd)/pr_results -s $(pwd)/pr -b /opt/benchmarks/build -o ~/data.osm.pbf -g ~/gps_traces.csv
           sudo umount /opt/benchmarks
       - run: ls -R /opt/benchmarks
-      # - name: Run Base Benchmarks
-      #   run: |
-      #     sudo mount -t tmpfs -o size=4g none /opt/benchmarks
-      #     cp -rf base/build /opt/benchmarks/build
-      #     mkdir -p /opt/benchmarks/test
-      #     cp -rf base/test/data /opt/benchmarks/test/data
-      #     cp -rf base/profiles /opt/benchmarks/profiles
-      #     ls -R /opt/benchmarks
-      #     # we intentionally use scripts from PR branch to be able to update them and see results in the same PR
-      #     ./pr/scripts/ci/run_benchmarks.sh -f /opt/benchmarks -r $(pwd)/base_results -s $(pwd)/pr -b /opt/benchmarks/build -o ~/data.osm.pbf -g ~/gps_traces.csv
-      #     sudo umount /opt/benchmarks
-      # - name: Post Benchmark Results
-      #   run: |
-      #     python3 pr/scripts/ci/post_benchmark_results.py base_results pr_results
+      - name: Run Base Benchmarks
+        run: |
+          sudo mount -t tmpfs -o size=4g none /opt/benchmarks
+          cp -rf base/build /opt/benchmarks/build
+          mkdir -p /opt/benchmarks/test
+          cp -rf base/test/data /opt/benchmarks/test/data
+          cp -rf base/profiles /opt/benchmarks/profiles
+          ls -R /opt/benchmarks
+          # we intentionally use scripts from PR branch to be able to update them and see results in the same PR
+          ./pr/scripts/ci/run_benchmarks.sh -f /opt/benchmarks -r $(pwd)/base_results -s $(pwd)/pr -b /opt/benchmarks/build -o ~/data.osm.pbf -g ~/gps_traces.csv
+          sudo umount /opt/benchmarks
+      - name: Post Benchmark Results
+        run: |
+          python3 pr/scripts/ci/post_benchmark_results.py base_results pr_results
       - name: Show CCache statistics
         run: |
           ccache -p