--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -743,6 +743,11 @@ jobs:
           mkdir -p /opt/benchmarks/test
           cp -rf base/test/data /opt/benchmarks/test/data
           cp -rf base/profiles /opt/benchmarks/profiles
+
+          # TODO: remove it when base branch will have this file at needed location
+          if [ ! -f /opt/benchmarks/test/data/portugal_to_korea.json ]; then
+            cp base/src/benchmarks/portugal_to_korea.json /opt/benchmarks/test/data/portugal_to_korea.json
+          fi
           ls -R /opt/benchmarks
           # we intentionally use scripts from PR branch to be able to update them and see results in the same PR
           ./pr/scripts/ci/run_benchmarks.sh -f /opt/benchmarks -r $(pwd)/base_results -s $(pwd)/pr -b /opt/benchmarks/build -o ~/data.osm.pbf -g ~/gps_traces.csv