Via using `tmpfs` (i.e. RAM disk) we avoid impact of disc IO when running benchmarks, i.e. make them a bit more stable (not stable enough yet though).



<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1086.68<br/>plain u32: 1060.68<br/>aliased double: 961.416<br/>plain double: 958.677 | aliased u32: 1091.17<br/>plain u32: 1085.88<br/>aliased double: 958.248<br/>plain double: 953.365 |
| e2e_match_ch | Ops: 43.29 ± 0.11 ops/s. Best: 43.15 ops/s<br/>Total: 3026.17ms ± 8.45ms. Best: 3007.44ms<br/>Min time: 2.24ms ± 0.04ms<br/>Mean time: 23.10ms ± 0.07ms<br/>Median time: 17.03ms ± 0.07ms<br/>95th percentile: 78.39ms ± 0.19ms<br/>99th percentile: 95.46ms ± 0.41ms<br/>Max time: 99.59ms ± 0.18ms | Ops: 43.63 ± 0.11 ops/s. Best: 43.46 ops/s<br/>Total: 3002.20ms ± 7.96ms. Best: 2987.51ms<br/>Min time: 2.21ms ± 0.06ms<br/>Mean time: 22.92ms ± 0.06ms<br/>Median time: 16.77ms ± 0.09ms<br/>95th percentile: 78.84ms ± 0.24ms<br/>99th percentile: 95.39ms ± 0.42ms<br/>Max time: 99.58ms ± 0.24ms |
| e2e_match_mld | Ops: 62.73 ± 0.07 ops/s. Best: 62.61 ops/s<br/>Total: 2088.22ms ± 2.23ms. Best: 2085.06ms<br/>Min time: 1.83ms ± 0.03ms<br/>Mean time: 15.94ms ± 0.02ms<br/>Median time: 8.52ms ± 0.07ms<br/>95th percentile: 53.30ms ± 0.14ms<br/>99th percentile: 61.31ms ± 0.29ms<br/>Max time: 71.16ms ± 0.46ms | Ops: 61.88 ± 0.13 ops/s. Best: 61.75 ops/s<br/>Total: 2117.23ms ± 4.13ms. Best: 2107.71ms<br/>Min time: 1.84ms ± 0.05ms<br/>Mean time: 16.16ms ± 0.03ms<br/>Median time: 8.61ms ± 0.06ms<br/>95th percentile: 53.84ms ± 0.13ms<br/>99th percentile: 62.27ms ± 0.17ms<br/>Max time: 72.23ms ± 0.10ms |
| e2e_nearest_ch | Ops: 809.69 ± 4.92 ops/s. Best: 798.58 ops/s<br/>Total: 1234.91ms ± 8.09ms. Best: 1227.26ms<br/>Min time: 1.05ms ± 0.01ms<br/>Mean time: 1.24ms ± 0.01ms<br/>Median time: 1.15ms ± 0.01ms<br/>95th percentile: 1.63ms ± 0.01ms<br/>99th percentile: 1.68ms ± 0.01ms<br/>Max time: 4.11ms ± 2.38ms | Ops: 795.44 ± 5.66 ops/s. Best: 781.62 ops/s<br/>Total: 1257.16ms ± 9.27ms. Best: 1247.84ms<br/>Min time: 1.07ms ± 0.01ms<br/>Mean time: 1.26ms ± 0.01ms<br/>Median time: 1.17ms ± 0.01ms<br/>95th percentile: 1.66ms ± 0.01ms<br/>99th percentile: 1.71ms ± 0.02ms<br/>Max time: 4.50ms ± 2.29ms |
| e2e_nearest_mld | Ops: 797.19 ± 5.00 ops/s. Best: 785.34 ops/s<br/>Total: 1254.36ms ± 8.27ms. Best: 1245.16ms<br/>Min time: 1.06ms ± 0.01ms<br/>Mean time: 1.25ms ± 0.01ms<br/>Median time: 1.17ms ± 0.01ms<br/>95th percentile: 1.66ms ± 0.01ms<br/>99th percentile: 1.71ms ± 0.02ms<br/>Max time: 4.23ms ± 2.52ms | Ops: 797.61 ± 6.27 ops/s. Best: 786.17 ops/s<br/>Total: 1253.82ms ± 10.78ms. Best: 1237.25ms<br/>Min time: 1.06ms ± 0.00ms<br/>Mean time: 1.25ms ± 0.01ms<br/>Median time: 1.16ms ± 0.01ms<br/>95th percentile: 1.65ms ± 0.01ms<br/>99th percentile: 1.71ms ± 0.01ms<br/>Max time: 4.29ms ± 2.55ms |
| e2e_route_ch | Ops: 337.00 ± 2.10 ops/s. Best: 334.01 ops/s<br/>Total: 2967.78ms ± 18.35ms. Best: 2945.88ms<br/>Min time: 1.29ms ± 0.01ms<br/>Mean time: 2.97ms ± 0.02ms<br/>Median time: 2.99ms ± 0.02ms<br/>95th percentile: 3.92ms ± 0.03ms<br/>99th percentile: 4.31ms ± 0.06ms<br/>Max time: 6.64ms ± 1.81ms | Ops: 337.07 ± 1.24 ops/s. Best: 334.35 ops/s<br/>Total: 2966.70ms ± 11.58ms. Best: 2952.26ms<br/>Min time: 1.30ms ± 0.02ms<br/>Mean time: 2.97ms ± 0.01ms<br/>Median time: 2.99ms ± 0.02ms<br/>95th percentile: 3.91ms ± 0.02ms<br/>99th percentile: 4.35ms ± 0.03ms<br/>Max time: 7.25ms ± 2.03ms |
| e2e_route_mld | Ops: 282.27 ± 1.99 ops/s. Best: 278.37 ops/s<br/>Total: 3543.30ms ± 26.85ms. Best: 3508.52ms<br/>Min time: 1.30ms ± 0.02ms<br/>Mean time: 3.54ms ± 0.03ms<br/>Median time: 3.58ms ± 0.04ms<br/>95th percentile: 4.83ms ± 0.03ms<br/>99th percentile: 5.31ms ± 0.04ms<br/>Max time: 7.39ms ± 1.55ms | Ops: 281.02 ± 2.62 ops/s. Best: 275.06 ops/s<br/>Total: 3560.31ms ± 33.58ms. Best: 3519.86ms<br/>Min time: 1.32ms ± 0.01ms<br/>Mean time: 3.56ms ± 0.03ms<br/>Median time: 3.60ms ± 0.04ms<br/>95th percentile: 4.84ms ± 0.06ms<br/>99th percentile: 5.33ms ± 0.08ms<br/>Max time: 7.76ms ± 1.50ms |
| e2e_table_ch | Ops: 302.59 ± 1.00 ops/s. Best: 301.06 ops/s<br/>Total: 3304.38ms ± 10.97ms. Best: 3290.93ms<br/>Min time: 1.70ms ± 0.02ms<br/>Mean time: 3.30ms ± 0.01ms<br/>Median time: 3.31ms ± 0.02ms<br/>95th percentile: 4.56ms ± 0.02ms<br/>99th percentile: 4.89ms ± 0.05ms<br/>Max time: 7.73ms ± 2.61ms | Ops: 304.47 ± 1.13 ops/s. Best: 302.64 ops/s<br/>Total: 3284.58ms ± 12.48ms. Best: 3262.49ms<br/>Min time: 1.75ms ± 0.01ms<br/>Mean time: 3.28ms ± 0.01ms<br/>Median time: 3.29ms ± 0.03ms<br/>95th percentile: 4.50ms ± 0.03ms<br/>99th percentile: 4.86ms ± 0.05ms<br/>Max time: 7.69ms ± 2.58ms |
| e2e_table_mld | Ops: 109.25 ± 0.14 ops/s. Best: 109.03 ops/s<br/>Total: 9153.40ms ± 11.81ms. Best: 9132.76ms<br/>Min time: 3.70ms ± 0.03ms<br/>Mean time: 9.15ms ± 0.01ms<br/>Median time: 9.12ms ± 0.05ms<br/>95th percentile: 13.99ms ± 0.03ms<br/>99th percentile: 14.93ms ± 0.12ms<br/>Max time: 17.62ms ± 1.57ms | Ops: 109.26 ± 0.09 ops/s. Best: 109.07 ops/s<br/>Total: 9152.29ms ± 8.56ms. Best: 9137.62ms<br/>Min time: 3.68ms ± 0.03ms<br/>Mean time: 9.15ms ± 0.01ms<br/>Median time: 9.12ms ± 0.01ms<br/>95th percentile: 13.99ms ± 0.02ms<br/>99th percentile: 14.92ms ± 0.06ms<br/>Max time: 16.91ms ± 1.35ms |
| e2e_trip_ch | Ops: 95.07 ± 0.24 ops/s. Best: 94.74 ops/s<br/>Total: 10517.13ms ± 26.37ms. Best: 10468.19ms<br/>Min time: 1.54ms ± 0.13ms<br/>Mean time: 10.52ms ± 0.03ms<br/>Median time: 10.00ms ± 0.02ms<br/>95th percentile: 18.58ms ± 0.11ms<br/>99th percentile: 20.32ms ± 0.08ms<br/>Max time: 21.83ms ± 0.29ms | Ops: 95.05 ± 0.31 ops/s. Best: 94.60 ops/s<br/>Total: 10520.48ms ± 36.60ms. Best: 10448.40ms<br/>Min time: 1.62ms ± 0.24ms<br/>Mean time: 10.52ms ± 0.04ms<br/>Median time: 9.99ms ± 0.02ms<br/>95th percentile: 18.56ms ± 0.05ms<br/>99th percentile: 20.18ms ± 0.12ms<br/>Max time: 21.99ms ± 0.16ms |
| e2e_trip_mld | Ops: 57.21 ± 0.15 ops/s. Best: 56.89 ops/s<br/>Total: 17478.10ms ± 47.96ms. Best: 17406.23ms<br/>Min time: 1.75ms ± 0.25ms<br/>Mean time: 17.48ms ± 0.05ms<br/>Median time: 17.08ms ± 0.06ms<br/>95th percentile: 28.60ms ± 0.06ms<br/>99th percentile: 30.46ms ± 0.20ms<br/>Max time: 33.05ms ± 0.73ms | Ops: 56.23 ± 0.41 ops/s. Best: 55.68 ops/s<br/>Total: 17791.47ms ± 125.57ms. Best: 17556.82ms<br/>Min time: 1.87ms ± 0.14ms<br/>Mean time: 17.79ms ± 0.13ms<br/>Median time: 17.40ms ± 0.16ms<br/>95th percentile: 29.01ms ± 0.28ms<br/>99th percentile: 31.08ms ± 0.30ms<br/>Max time: 34.08ms ± 1.23ms |
| json-render | String: 6.94545ms<br/>Stringstream: 10.6506ms<br/>Vector: 7.01163ms | String: 6.67249ms<br/>Stringstream: 10.4753ms<br/>Vector: 6.89838ms |
| match_ch | Default radius: <br/>4.55493ms/req at 82 coordinate<br/>0.055548ms/coordinate<br/>Radius 10m: <br/>15.9693ms/req at 82 coordinate<br/>0.194748ms/coordinate | Default radius: <br/>4.64392ms/req at 82 coordinate<br/>0.0566332ms/coordinate<br/>Radius 10m: <br/>16.101ms/req at 82 coordinate<br/>0.196354ms/coordinate |
| match_mld | Default radius: <br/>3.06623ms/req at 82 coordinate<br/>0.0373931ms/coordinate<br/>Radius 10m: <br/>11.4996ms/req at 82 coordinate<br/>0.140238ms/coordinate | Default radius: <br/>3.1139ms/req at 82 coordinate<br/>0.0379744ms/coordinate<br/>Radius 10m: <br/>11.7394ms/req at 82 coordinate<br/>0.143164ms/coordinate |
| osrm_contract | Time: 96.83s Peak RAM: 196.22MB | Time: 95.21s Peak RAM: 195.63MB |
| osrm_customize | Time: 1.34s Peak RAM: 116.83MB | Time: 1.36s Peak RAM: 116.67MB |
| osrm_extract | Time: 12.12s Peak RAM: 409.35MB | Time: 12.14s Peak RAM: 407.56MB |
| osrm_partition | Time: 2.04s Peak RAM: 136.42MB | Time: 2.02s Peak RAM: 132.78MB |
| packedvector | random write:<br/>std::vector 9898.25 ms<br/>util::packed_vector 82051.2 ms<br/>slowdown: 8.28947<br/>random read:<br/>std::vector 8451.95 ms<br/>util::packed_vector 33467.7 ms<br/>slowdown: 3.95976 | random write:<br/>std::vector 11263.4 ms<br/>util::packed_vector 78249.4 ms<br/>slowdown: 6.94723<br/>random read:<br/>std::vector 11232.6 ms<br/>util::packed_vector 32529.3 ms<br/>slowdown: 2.89598 |
| random_match_ch | 500 matches, default radius<br/>ops: 209.46 ± 0.75 ops/s. best: 207.80ops/s.<br/>total: 272.13 ± 1.00ms. best: 271.06ms.<br/>avg: 4.77 ± 0.02ms<br/>min: 0.16 ± 0.01ms<br/>max: 24.05 ± 0.11ms<br/>p99: 24.05 ± 0.11ms<br/><br/>500 matches, radius=10<br/>ops: 62.24 ± 0.13 ops/s. best: 62.08ops/s.<br/>total: 1028.23 ± 2.09ms. best: 1024.10ms.<br/>avg: 16.07 ± 0.03ms<br/>min: 0.15 ± 0.00ms<br/>max: 229.42 ± 0.85ms<br/>p99: 229.42 ± 0.85ms<br/><br/>500 matches, radius=20<br/>ops: 14.76 ± 0.01 ops/s. best: 14.74ops/s.<br/>total: 4403.50 ± 4.13ms. best: 4396.97ms.<br/>avg: 67.75 ± 0.06ms<br/>min: 0.34 ± 0.01ms<br/>max: 1152.76 ± 2.78ms<br/>p99: 1152.76 ± 2.78ms | 500 matches, default radius<br/>ops: 211.93 ± 1.70 ops/s. best: 207.98ops/s.<br/>total: 268.99 ± 2.19ms. best: 267.07ms.<br/>avg: 4.72 ± 0.04ms<br/>min: 0.16 ± 0.01ms<br/>max: 25.25 ± 0.12ms<br/>p99: 25.25 ± 0.12ms<br/><br/>500 matches, radius=10<br/>ops: 62.52 ± 0.38 ops/s. best: 61.62ops/s.<br/>total: 1023.68 ± 6.30ms. best: 1017.15ms.<br/>avg: 15.99 ± 0.10ms<br/>min: 0.15 ± 0.01ms<br/>max: 245.97 ± 0.60ms<br/>p99: 245.97 ± 0.60ms<br/><br/>500 matches, radius=20<br/>ops: 14.82 ± 0.02 ops/s. best: 14.80ops/s.<br/>total: 4384.52 ± 5.67ms. best: 4373.03ms.<br/>avg: 67.45 ± 0.09ms<br/>min: 0.29 ± 0.00ms<br/>max: 1232.94 ± 3.58ms<br/>p99: 1232.94 ± 3.58ms |
| random_match_mld | 500 matches, default radius<br/>ops: 302.95 ± 1.40 ops/s. best: 299.51ops/s.<br/>total: 188.15 ± 0.88ms. best: 187.41ms.<br/>avg: 3.30 ± 0.02ms<br/>min: 0.14 ± 0.00ms<br/>max: 19.54 ± 0.04ms<br/>p99: 19.54 ± 0.04ms<br/><br/>500 matches, radius=10<br/>ops: 106.95 ± 0.05 ops/s. best: 106.87ops/s.<br/>total: 598.43 ± 0.27ms. best: 597.98ms.<br/>avg: 9.35 ± 0.00ms<br/>min: 0.14 ± 0.00ms<br/>max: 114.31 ± 0.24ms<br/>p99: 114.31 ± 0.24ms<br/><br/>500 matches, radius=20<br/>ops: 21.41 ± 0.04 ops/s. best: 21.33ops/s.<br/>total: 3036.54 ± 5.19ms. best: 3030.84ms.<br/>avg: 46.72 ± 0.08ms<br/>min: 0.20 ± 0.00ms<br/>max: 599.39 ± 1.95ms<br/>p99: 599.39 ± 1.95ms | 500 matches, default radius<br/>ops: 304.45 ± 1.32 ops/s. best: 301.29ops/s.<br/>total: 187.23 ± 0.84ms. best: 186.49ms.<br/>avg: 3.28 ± 0.01ms<br/>min: 0.14 ± 0.00ms<br/>max: 19.38 ± 0.05ms<br/>p99: 19.38 ± 0.05ms<br/><br/>500 matches, radius=10<br/>ops: 107.59 ± 0.08 ops/s. best: 107.45ops/s.<br/>total: 594.86 ± 0.46ms. best: 594.24ms.<br/>avg: 9.29 ± 0.01ms<br/>min: 0.14 ± 0.00ms<br/>max: 113.44 ± 0.13ms<br/>p99: 113.44 ± 0.13ms<br/><br/>500 matches, radius=20<br/>ops: 21.47 ± 0.04 ops/s. best: 21.41ops/s.<br/>total: 3026.83 ± 5.03ms. best: 3020.48ms.<br/>avg: 46.57 ± 0.08ms<br/>min: 0.20 ± 0.00ms<br/>max: 598.72 ± 1.13ms<br/>p99: 598.72 ± 1.13ms |
| random_nearest_ch | 10000 nearest, number_of_results=1<br/>ops: 22622.86 ± 59.25 ops/s. best: 22498.50ops/s.<br/>total: 442.03 ± 1.18ms. best: 440.49ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.16 ± 0.04ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17627.82 ± 28.55 ops/s. best: 17574.59ops/s.<br/>total: 567.29 ± 0.91ms. best: 566.26ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.13 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 13989.21 ± 29.95 ops/s. best: 13936.52ops/s.<br/>total: 714.84 ± 1.53ms. best: 712.35ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.14 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 22504.74 ± 53.37 ops/s. best: 22403.00ops/s.<br/>total: 444.35 ± 1.06ms. best: 442.85ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.04ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17301.27 ± 37.37 ops/s. best: 17253.43ops/s.<br/>total: 578.00 ± 1.25ms. best: 576.18ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.13 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 13591.78 ± 5.74 ops/s. best: 13581.84ops/s.<br/>total: 735.74 ± 0.31ms. best: 735.23ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.14 ± 0.00ms |
| random_nearest_mld | 10000 nearest, number_of_results=1<br/>ops: 22357.07 ± 56.39 ops/s. best: 22271.49ops/s.<br/>total: 447.29 ± 1.13ms. best: 445.44ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.04ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17573.60 ± 59.52 ops/s. best: 17453.64ops/s.<br/>total: 569.04 ± 1.93ms. best: 566.49ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.13 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 14137.99 ± 22.66 ops/s. best: 14100.21ops/s.<br/>total: 707.32 ± 1.14ms. best: 705.62ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.17 ± 0.01ms<br/>p99: 0.14 ± 0.00ms | 10000 nearest, number_of_results=1<br/>ops: 22264.04 ± 89.25 ops/s. best: 22077.20ops/s.<br/>total: 449.16 ± 1.81ms. best: 447.27ms.<br/>avg: 0.04 ± 0.00ms<br/>min: 0.01 ± 0.00ms<br/>max: 0.17 ± 0.04ms<br/>p99: 0.11 ± 0.00ms<br/><br/>10000 nearest, number_of_results=5<br/>ops: 17228.47 ± 33.20 ops/s. best: 17176.27ops/s.<br/>total: 580.44 ± 1.12ms. best: 578.65ms.<br/>avg: 0.06 ± 0.00ms<br/>min: 0.02 ± 0.00ms<br/>max: 0.15 ± 0.00ms<br/>p99: 0.13 ± 0.00ms<br/><br/>10000 nearest, number_of_results=10<br/>ops: 13554.10 ± 46.52 ops/s. best: 13469.60ops/s.<br/>total: 737.80 ± 2.54ms. best: 734.97ms.<br/>avg: 0.07 ± 0.00ms<br/>min: 0.03 ± 0.00ms<br/>max: 0.18 ± 0.01ms<br/>p99: 0.14 ± 0.00ms |
| random_route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 470.17 ± 3.20 ops/s. best: 465.17ops/s.<br/>total: 2092.97 ± 14.22ms. best: 2066.23ms.<br/>avg: 2.13 ± 0.01ms<br/>min: 0.32 ± 0.00ms<br/>max: 3.64 ± 0.16ms<br/>p99: 3.11 ± 0.06ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 540.37 ± 1.43 ops/s. best: 539.13ops/s.<br/>total: 1850.60 ± 4.89ms. best: 1838.79ms.<br/>avg: 1.85 ± 0.00ms<br/>min: 0.06 ± 0.00ms<br/>max: 4.49 ± 0.12ms<br/>p99: 4.03 ± 0.03ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 950.42 ± 4.03 ops/s. best: 944.18ops/s.<br/>total: 1035.36 ± 4.38ms. best: 1027.56ms.<br/>avg: 1.05 ± 0.00ms<br/>min: 0.24 ± 0.00ms<br/>max: 1.75 ± 0.04ms<br/>p99: 1.50 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 990.71 ± 31.04 ops/s. best: 920.39ops/s.<br/>total: 1010.84 ± 33.73ms. best: 975.20ms.<br/>avg: 1.01 ± 0.03ms<br/>min: 0.05 ± 0.00ms<br/>max: 2.96 ± 0.08ms<br/>p99: 2.39 ± 0.08ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 474.82 ± 2.31 ops/s. best: 470.35ops/s.<br/>total: 2072.43 ± 10.11ms. best: 2058.38ms.<br/>avg: 2.11 ± 0.01ms<br/>min: 0.32 ± 0.00ms<br/>max: 3.65 ± 0.15ms<br/>p99: 3.06 ± 0.02ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 567.39 ± 3.88 ops/s. best: 560.11ops/s.<br/>total: 1762.57 ± 12.12ms. best: 1749.36ms.<br/>avg: 1.76 ± 0.01ms<br/>min: 0.06 ± 0.00ms<br/>max: 4.79 ± 0.03ms<br/>p99: 3.97 ± 0.05ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 955.18 ± 3.16 ops/s. best: 949.40ops/s.<br/>total: 1030.19 ± 3.52ms. best: 1026.09ms.<br/>avg: 1.05 ± 0.00ms<br/>min: 0.26 ± 0.00ms<br/>max: 1.61 ± 0.02ms<br/>p99: 1.48 ± 0.01ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 1086.94 ± 11.66 ops/s. best: 1069.46ops/s.<br/>total: 920.17 ± 9.98ms. best: 909.15ms.<br/>avg: 0.92 ± 0.01ms<br/>min: 0.05 ± 0.01ms<br/>max: 2.94 ± 0.05ms<br/>p99: 2.29 ± 0.03ms |
| random_route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 237.40 ± 0.77 ops/s. best: 235.87ops/s.<br/>total: 4144.91 ± 15.19ms. best: 4127.32ms.<br/>avg: 4.21 ± 0.02ms<br/>min: 0.32 ± 0.01ms<br/>max: 9.14 ± 0.22ms<br/>p99: 7.05 ± 0.05ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 230.20 ± 0.99 ops/s. best: 228.80ops/s.<br/>total: 4344.24 ± 18.62ms. best: 4314.08ms.<br/>avg: 4.34 ± 0.02ms<br/>min: 0.06 ± 0.00ms<br/>max: 9.94 ± 0.23ms<br/>p99: 8.79 ± 0.05ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 320.13 ± 1.85 ops/s. best: 317.17ops/s.<br/>total: 3073.92 ± 17.99ms. best: 3054.18ms.<br/>avg: 3.12 ± 0.02ms<br/>min: 0.29 ± 0.00ms<br/>max: 7.44 ± 0.07ms<br/>p99: 5.33 ± 0.08ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 291.90 ± 0.93 ops/s. best: 289.99ops/s.<br/>total: 3425.83 ± 11.12ms. best: 3410.51ms.<br/>avg: 3.43 ± 0.01ms<br/>min: 0.04 ± 0.00ms<br/>max: 7.72 ± 0.32ms<br/>p99: 6.73 ± 0.06ms | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>ops: 239.49 ± 1.24 ops/s. best: 237.84ops/s.<br/>total: 4108.84 ± 21.21ms. best: 4074.68ms.<br/>avg: 4.18 ± 0.02ms<br/>min: 0.32 ± 0.00ms<br/>max: 9.01 ± 0.13ms<br/>p99: 7.00 ± 0.06ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>ops: 232.62 ± 2.14 ops/s. best: 229.06ops/s.<br/>total: 4299.27 ± 39.79ms. best: 4260.03ms.<br/>avg: 4.30 ± 0.04ms<br/>min: 0.06 ± 0.00ms<br/>max: 9.80 ± 0.25ms<br/>p99: 8.58 ± 0.12ms<br/><br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>ops: 321.77 ± 3.25 ops/s. best: 314.94ops/s.<br/>total: 3058.43 ± 31.08ms. best: 3019.23ms.<br/>avg: 3.11 ± 0.03ms<br/>min: 0.29 ± 0.00ms<br/>max: 7.52 ± 0.24ms<br/>p99: 5.36 ± 0.14ms<br/><br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>ops: 292.64 ± 2.17 ops/s. best: 288.36ops/s.<br/>total: 3417.41 ± 25.54ms. best: 3382.76ms.<br/>avg: 3.42 ± 0.03ms<br/>min: 0.05 ± 0.00ms<br/>max: 7.78 ± 0.16ms<br/>p99: 6.65 ± 0.08ms |
| random_table_ch | 250 tables, 3 coordinates<br/>ops: 1421.99 ± 8.43 ops/s. best: 1403.04ops/s.<br/>total: 175.82 ± 1.05ms. best: 174.69ms.<br/>avg: 0.70 ± 0.00ms<br/>min: 0.52 ± 0.00ms<br/>max: 1.06 ± 0.20ms<br/>p99: 0.89 ± 0.05ms<br/><br/>250 tables, 25 coordinates<br/>ops: 167.98 ± 0.19 ops/s. best: 167.59ops/s.<br/>total: 1488.29 ± 1.66ms. best: 1486.53ms.<br/>avg: 5.95 ± 0.01ms<br/>min: 5.23 ± 0.00ms<br/>max: 6.56 ± 0.03ms<br/>p99: 6.44 ± 0.02ms<br/><br/>250 tables, 50 coordinates<br/>ops: 82.68 ± 0.16 ops/s. best: 82.32ops/s.<br/>total: 3023.66 ± 5.92ms. best: 3017.10ms.<br/>avg: 12.09 ± 0.02ms<br/>min: 11.09 ± 0.06ms<br/>max: 13.38 ± 0.21ms<br/>p99: 13.03 ± 0.10ms | 250 tables, 3 coordinates<br/>ops: 1442.16 ± 25.52 ops/s. best: 1381.31ops/s.<br/>total: 173.43 ± 3.24ms. best: 170.64ms.<br/>avg: 0.69 ± 0.01ms<br/>min: 0.53 ± 0.00ms<br/>max: 1.01 ± 0.22ms<br/>p99: 0.88 ± 0.09ms<br/><br/>250 tables, 25 coordinates<br/>ops: 170.83 ± 1.08 ops/s. best: 168.29ops/s.<br/>total: 1463.54 ± 9.31ms. best: 1454.49ms.<br/>avg: 5.85 ± 0.04ms<br/>min: 5.15 ± 0.01ms<br/>max: 6.75 ± 0.64ms<br/>p99: 6.66 ± 0.56ms<br/><br/>250 tables, 50 coordinates<br/>ops: 84.41 ± 0.23 ops/s. best: 84.07ops/s.<br/>total: 2961.89 ± 8.08ms. best: 2949.54ms.<br/>avg: 11.85 ± 0.03ms<br/>min: 10.96 ± 0.12ms<br/>max: 12.96 ± 0.34ms<br/>p99: 12.70 ± 0.13ms |
…